requirements.txt
langchain
langchain_groq
langchain_core
langchain_community
langchain-text-splitters
langchain-openai==0.1.7
openai
python-dotenv
ipykernel
pypdf
chromadb
faiss-cpu
beautifulsoup4
pandas
defusedxml
scikit-learn
wikipedia
langgraph
matplotlib
xmltodict
pydantic==1.10.13
streamlit==1.35.0